% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_scale_length.R
\name{filter_scale_length}
\alias{filter_scale_length}
\title{filter_scale_length}
\usage{
filter_scale_length(data, scale_names, scale_values, min_scale_length = 4)
}
\arguments{
\item{data}{A data set object.}

\item{scale_names}{A column containing scale names.}

\item{scale_values}{A column containing scale values.}

\item{min_scale_length}{The threshold below which to remove ordered ratings scales.}
}
\value{
The data set with ordered ratings scales below the threshold removed.
}
\description{
Filters ordered ratings scales with a minimum length of unique values.
}
\examples{
data(toydata)
# The default is to filter ordered ratings scales with a minimum length of 4 unique values
toydata_l <- filter_scale_length(toydata_l, scale_names = att_name, scale_values = att_val)
# To change the threshold, set min_scale_length to the desired value
toydata_l <- filter_scale_length(toydata_l, scale_names = att_name, scale_values = att_val, min_scale_length = 5)

}
