% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarize_assoc.R
\name{polarize_assoc}
\alias{polarize_assoc}
\title{polarize_assoc}
\usage{
polarize_assoc(
  data,
  value_1,
  value_2,
  r_or_r2,
  by = NULL,
  ids = NULL,
  probs = NULL,
  strata = NULL,
  fpc = NULL,
  weights = NULL,
  nest = FALSE
)
}
\arguments{
\item{data}{A data set object}

\item{value_1}{A column containing values on some variable or set of variables, such as attitude item responses.}

\item{value_2}{A column containing values on a second variable or set of variables (e.g., political party affiliation).}

\item{r_or_r2}{A string specifying the measure of association. See Details below.}

\item{by}{Optional groups to nest observations by (e.g., attitude item, survey wave, country).}

\item{ids}{Variables specifying cluster ids from largest level to smallest level (leaving the argument empty, NULL, 1, or 0 indicate no clusters).}

\item{probs}{Variables specifying cluster sampling probabilities.}

\item{strata}{Variables specifying strata.}

\item{fpc}{Variables specifying a finite population correct, see
\code{\link[survey]{svydesign}} for more details.}

\item{weights}{Variables specifying weights (inverse of probability).}

\item{nest}{If \code{TRUE}, relabel cluster ids to enforce nesting within strata.}
}
\value{
A data frame object containing the measure of association between \code{value_1} and \code{value_2} for any groups in \code{by}.
}
\description{
Calculates associational measures of polarization
}
\details{
This function implements two prominent associational measures of polarization. It is designed around the \code{survey} package, allowing the incorporation of complex survey design features. It is useful when you want to calculate associational measures of polarization over a large number of attitude items. This was previously less convenient in the \code{survey} package, which requires the user to specify variables manually. Pass columns containing grouping information (such as variable names) to the \code{by} argument, and \code{polarize_assoc} will automatically nest the data and apply functions related to the \code{survey} package.

When \code{r_or_r2} is set to "r", \code{polarize_assoc} transforms the underlying covariance matrix calculated by \code{survey::svyvar} into Pearson correlation coefficients using the \code{cov2cor} function. This approach is recommended by the \code{survey} package author, Thomas Lumley (see {https://stackoverflow.com/questions/34418822/pearson-correlation-coefficient-in-rs-survey-package#41031088}{this} discussion on Stack Overflow).

When \code{r_or_r2} is set to "r2", \code{polarize_assoc} fits OLS regression models using \code{survey::svyglm}. The column supplied to \code{value_1} will be used as the dependent variable and the column supplied to \code{value_2} will be the independent variable. The direction of this relationship makes no difference in practice, however. The R-squared and adjusted R-squared are extracted from each model and returned as the output, giving the square of the correlation between observed and predicted outcomes. These statistics are similar to squaring the Pearson correlation between two variables, only the regression framework allows for the incorporation of categorical predictors, such as unordered party values (see Caldwell, Cohen, and Vivyan 2023).
}
\examples{
data(toydata)
# Partisan polarization using Pearson correlation between attitude item and ordinal party affiliation variable
party_pol <- polarize_assoc(data = toydata, value_1 = att_val, value_2 = party_ord, r_or_r2 = "r", by = c(att_name, group))

# Partisan polarization using unordered party categories and no groups
multiparty_pol <- polarize_assoc(data = toydata, value_1 = att_val, value_2 = party_cat, r_or_r2 = "r2")

# Ideological polarization using the Pearson correlation between attitude pairs and no groups
paired_toydata <- spread_pairs(data = toydata, name_key = att_name, value_key = att_val, other_keys = c(id, group))
ideology_pol <- polarize_assoc(data = paired_toydata, value_1 = att_val1, value_2 = att_val2, r_or_r2 = "r")

# Ideological polarization grouping R-squared statistics by attitude pair
ideology_pol2 <- polarize_assoc(data = paired_toydata, value_1 = att_val1, value_2 = att_val2, r_or_r2 = "r2", by = c(att_name1, att_name2))

}
\references{
Caldwell, D.,  Cohen, C. and Vivyan, N. (2023). Long-Run Trends in Political Polarization of Climate Policy-Relevant Attitudes Across Countries. \emph{Working Paper}.
}
\seealso{
\code{\link[survey]{svyvar}}, \code{\link[survey]{svyglm}}
}
