% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarize_distr.R
\name{polarize_distr}
\alias{polarize_distr}
\title{polarize_distr}
\usage{
polarize_distr(
  data,
  value,
  measure,
  by = NULL,
  rescale_0_1 = FALSE,
  ids = NULL,
  probs = NULL,
  strata = NULL,
  fpc = NULL,
  weights = NULL,
  nest = FALSE
)
}
\arguments{
\item{data}{A data set object}

\item{value}{A column containing values on some variable or set of variables, such as attitude item responses.}

\item{measure}{A string specifying the measure of distributional polarization. See Details below.}

\item{by}{A character vector of optional groups to nest observations by (e.g., survey wave, country, social group).}

\item{rescale_0_1}{Logical. Setting to TRUE normalizes values prior to summarizing their distribution.}

\item{ids}{Variables specifying cluster ids from largest level to smallest level (leaving the argument empty, NULL, 1, or 0 indicate no clusters).}

\item{probs}{Variables specifying cluster sampling probabilities.}

\item{strata}{Variables specifying strata.}

\item{fpc}{Variables specifying a finite population correct, see
\code{\link[survey]{svydesign}} for more details.}

\item{weights}{Variables specifying weights (inverse of probability).}

\item{nest}{If \code{TRUE}, relabel cluster ids to enforce nesting within strata.}
}
\value{
A data frame object containing the distributional measure applied to \code{value} for any groups in \code{by}.
}
\description{
Calculates distributional measures of polarization
}
\details{
This function is a one-stop-shop for distributional measures of polarization. It is designed around the \code{survey} package, allowing for the incorporation of complex survey design features. It is useful when you want to summarise a large number of attitude distributions at once. This was previously less convenient, as the \code{survey} package requires the user to specify variables manually. Pass columns containing grouping information (such as variable names) to the \code{by} argument, and \code{polarize_distr} will automatically nest the data and apply functions related to the \code{survey} package.

The latter is set by the \code{measure} argument, which currently accepts the following values:
\itemize{
 \item \code{"mean"}: Mean of the distribution, using \code{survey::svymean}
 \item \code{"median"}: Median of the distribution, using \code{survey::svyquantile}
\item \code{"iqr"}: Interquartile range of the distribution, using \code{survey::svyquantile}
\item \code{"var"}: Variance of the distribution, using \code{survey::svyvar}
\item \code{"std"}: Standard deviation of the distribution, using \code{sqrt(survey::svyvar)}
\item \code{"kurt"}: Kurtosis of the distribution, using the \code{svykurt} function implemented in this package
\item \code{"skew"}: Skewness of the distribution, using the \code{svyskew} function implemented in this package
\item \code{"extremism"}: Proportion of respondents who are extreme on the distribution, using the \code{svyextremism} function implemented in this package.
\item And these methods for estimating consensus and disagreement on ordered rating scales from the \code{agrmt} package, each of which are run on frequency vectors created using \code{survey::svytable}: \code{"agreement"}, \code{"polarization"}, \code{"Leik"}, \code{"consensus"}, \code{"entropy"}, \code{"BerryMielke"}, \code{"BlairLacy"}, \code{"Kvalseth"}, \code{"lsquared"}, \code{"dsquared"}, \code{"MRQ"}, \code{"concentration"}, \code{"dispersion"}, and \code{"Reardon"}. See \code{\link[agrmt]{agreement}} for more details.}
}
\examples{
data(toydata)
# Filter attitude items with a length of unique values above 3
filtered_toydata <- filter_scale_length(toydata, scale_names = att_name, scale_values = att_val)

# Describing the central tendency of distributions surrounding attitude items
att_means <- polarize_distr(data = filtered_toydata, value = att_val, measure = "mean", by = c("att_name", "group"), rescale_0_1 = TRUE)
att_medians <- polarize_distr(data = filtered_toydata, value = att_val, measure = "median", by = c("att_name", "group"), rescale_0_1 = TRUE)

# Describing the dispersion of distributions surrounding attitude items
att_var <- polarize_distr(data = filtered_toydata, value = att_val, measure = "var", by = c("att_name", "group"), rescale_0_1 = TRUE)
att_std <- polarize_distr(data = filtered_toydata, value = att_val, measure = "std", by = c("att_name", "group"), rescale_0_1 = TRUE)
att_iqr <- polarize_distr(data = filtered_toydata, value = att_val, measure = "iqr", by = c("att_name", "group"), rescale_0_1 = TRUE)
att_kurt <- polarize_distr(data = filtered_toydata, value = att_val, measure = "kurt", by = c("att_name", "group"))
att_skew <- polarize_distr(data = filtered_toydata, value = att_val, measure = "skew", by = c("att_name", "group"))
att_extremism <- polarize_distri(data = filtered_toydata, value = att_val, measure = "extremism", by = c("att_name", "group"))

# Using measures of ordinal disgreement from the agrmt package
att_agreement <- polarize_distr(data = filtered_toydata, value = att_val, measure = "agreement", by = c("att_name", "group"))
att_polarization <- polarize_distr(data = filtered_toydata, value = att_val, measure = "polarization", by = c("att_name", "group"))
att_Leik <- polarize_distr(data = filtered_toydata, value = att_val, measure = "Leik", by = c("att_name", "group"))
att_consensus <- polarize_distr(data = filtered_toydata, value = att_val, measure = "consensus", by = c("att_name", "group"))
att_entropy <- polarize_distr(data = filtered_toydata, value = att_val, measure = "entropy", by = c("att_name", "group"))
att_BerryMielke <- polarize_distr(data = filtered_toydata, value = att_val, measure = "BerryMielke", by = c("att_name", "group"))
att_BlairLacy <- polarize_distr(data = filtered_toydata, value = att_val, measure = "BlairLacy", by = c("att_name", "group"))
att_Kvalseth <- polarize_distr(data = filtered_toydata, value = att_val, measure = "Kvalseth", by = c("att_name", "group"))
att_lsquared <- polarize_distr(data = filtered_toydata, value = att_val, measure = "lsquared", by = c("att_name", "group"))
att_dsquared <- polarize_distr(data = filtered_toydata, value = att_val, measure = "dsquared", by = c("att_name", "group"))
att_MRQ <- polarize_distr(data = filtered_toydata, value = att_val, measure = "MRQ", by = c("att_name", "group"))
att_concentration <- polarize_distr(data = filtered_toydata, value = att_val, measure = "concentration", by = c("att_name", "group"))
att_dispersion <- polarize_distr(data = filtered_toydata, value = att_val, measure = "dispersion", by = c("att_name", "group"))
att_Reardon <- polarize_distr(data = filtered_toydata, value = att_val, measure = "Reardon", by = c("att_name", "group"))

}
\seealso{
\code{\link[survey]{svymean}}, \code{\link[survey]{svyquantile}}, \code{\link[survey]{svyvar}}, \code{\link[survey]{svytable}}, \code{\link[polaRized]{svykurt}}, \code{\link[polaRized]{svyskew}}, \code{\link[polaRized]{svyextremism}}, \code{\link[agrmt]{agreement}}, \code{\link[agrmt]{polarization}}, \code{\link[agrmt]{Leik}}, \code{\link[agrmt]{consensus}}, \code{\link[agrmt]{entropy}}, \code{\link[agrmt]{BerryMielke}}, \code{\link[agrmt]{BlairLacy}}, \code{\link[agrmt]{Kvalseth}}, \code{\link[agrmt]{lsquared}}, \code{\link[agrmt]{dsquared}}, \code{\link[agrmt]{MRQ}}, \code{\link[agrmt]{concentration}}, \code{\link[agrmt]{dispersion}}, \code{\link[agrmt]{Reardon}}
}
