% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polaRized-package.R
\docType{package}
\name{polaRized-package}
\alias{polaRized}
\alias{polaRized-package}
\title{polaRized: Estimating Political Attitude Polarization in Complex Survey Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A one-stop-shop for measures of attitude polarization commonly used in the political science literature. It makes it easy to summarize distributional and associational properties among potentially large numbers of attitude items in complex survey data. The package also provides useful functions for incorporating survey design features into the estimation of key statistics, such as skewness and kurtosis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dcaldwellphd/polaRized}
  \item Report bugs at \url{https://github.com/dcaldwellphd/polaRized/issues}
}

}
\author{
\strong{Maintainer}: David Caldwell \email{dcaldwellphd@gmail.com}

}
\keyword{internal}
