% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_pairs.R
\name{spread_pairs}
\alias{spread_pairs}
\title{spread_pairs}
\usage{
spread_pairs(data, name_key, value_key, other_keys = NULL)
}
\arguments{
\item{data}{A data set object}

\item{name_key}{A column containing name keys}

\item{value_key}{A column containing value keys}

\item{other_keys}{Other columns to join by and preserve in the output}
}
\value{
A data frame object with name and value keys spread across unique pairs of observations
}
\description{
Takes a name key and value key and spreads into unique pairs of observations
}
\examples{
data(toydata)
paired_items <- spread_pairs(data = toydata, name_key = att_name, value_key = att_val, other_keys = c(id, group))

}
