% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyextremism.R
\name{svyextremism}
\alias{svyextremism}
\title{svyextremism}
\usage{
svyextremism(
  formula,
  design,
  na.rm = FALSE,
  method = "logit",
  level = 0.95,
  df = degf(design)
)
}
\arguments{
\item{formula}{A formula (e.g., ~var) specifying the variable on which to estimate extremism.}

\item{design}{The \code{survey.design} or \code{svyrep.design} object.}

\item{na.rm}{Logical. Should cases with missing values be dropped?}

\item{method}{How to estimate confidence interval. See \code{\link[survey]{svyciprop}} for more details.}

\item{level}{Confidence level for interval.}

\item{df}{Denominator degrees of freedom. See \code{\link[survey]{svyciprop}} for more details.}
}
\value{
An object of class \code{svyciprop} giving the proportion and its confidence interval
}
\description{
Calculates the proportion of extremism on ordered ratings scales in complex survey data
}
\details{
This function wraps around the \code{svyciprop} function from the \code{survey} package. It is like calling \code{svyciprop(~I(variable == value), design)}, but \code{svyextremism} automatically classifies and measures the proportion of extreme values on ordered ratings scales with different lengths. If the scale has less than 10 (and more than 3) unique values, the function uses its minimum and maximum as extreme values. If the scale has 10 or 11 unique values, extreme values also include the second lowest and highest response categories. If the scale is a feeling thermometer with 100 or 101 unique response categories, the top 20 and bottom 20 unique values are classed as extreme. Calling \code{svyextremism} on any other scale length returns an error. The approach to measuring extremism implemented here is based on previous research into attitude polarization (Adams et al. 2012; Caldwell 2023; Cohen & Cohen 2021), and allows the \code{polarize_distr} function from this package to loop over heterogeneous rataing scales.
}
\note{
This function was written without the knowledge or endorsement of the \code{survey} package authors.
}
\examples{
library(survey)
data(toydata)
# Create survey design object
toydesign <- svydesign(data = toydata_w, ids = ~1, weights = ~weight)
# Print the proportion of extremism on a variable without specifying extreme values
svyextremism(~att5val, design = toydesign)

}
\references{
Adams, J., Green, J., and Milazzo, C. (2012b). Has the British Public Depolarized Along With Political Elites? An American Perspective on British Public Opinion. \emph{Comparative Political Studies}, 45(4):507–530.

Caldwell, D. (2023) \emph{Polarisation and Cultural Realignment in Britain, 2014-2019}. Doctoral thesis, Durham University.

Cohen, G. and Cohen, S. (2021). Depolarization, Repolarization and Redistributive Ideological Change in Britain, 1983–2016. \emph{British Journal of Political Science}, 51(3):1181–1202.
}
\seealso{
[`svyciprop()`][survey::svyciprop]
}
