% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyskew.R
\name{svyskew}
\alias{svyskew}
\title{svyskew}
\usage{
svyskew(formula, design, na.rm = FALSE)
}
\arguments{
\item{formula}{A formula (e.g., ~var1) specifying the variable on which to estimate skewness.}

\item{design}{The \code{survey.design} or \code{svyrep.design} object.}

\item{na.rm}{Logical. Should cases with missing values be dropped?}
}
\value{
An object of class \code{svrepstat} giving the skewness and its standard error.
}
\description{
Calculates standardized skewness with complex survey data
}
\details{
This function extends the \code{survey} package by calculating the skewness of a variable in a complex survey design. It writes skewness in terms of raw moments and transforms via \code{svycontrast}.
}
\note{
This function generalizes the approach to estimating skewness provided in the "Examples" section of the documentation for \code{survey::svycontrast}. It has been developed without the knowledge or endorsement of the \code{survey} package authors.
}
\examples{
library(survey)
data(toydata)
# Create survey design object
toydesign <- svydesign(data = toydata_w, ids = ~1, weights = ~weight)
# Print the standardized skewness coefficient of a variable
svyskew(formula = ~att5val, design = toydesign)

}
\seealso{
[`svycontrast()`][survey::svycontrast]
}
